LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/event_play_noautoplay-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: The operation was aborted.
spec reference

  

Harness Error (FAIL), message = The operation was aborted.

PASS audio events - play 
PASS calling play() on audio should trigger play event 
PASS video events - play 
PASS calling play() on video should trigger play event 
